Break the Outlet/Cover/Bulb/Cloud modules out into separate files
